2_image_normalisation/requirements_for_visualisation_n_evaluation_env.txt
affine==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anndata==0.9.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bleach==6.1.0
boto3==1.35.73
botocore==1.35.73
cached-property==1.5.2
cffi==1.17.1
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
comm==0.2.2
contextlib2==21.6.0
contourpy==1.1.1
ctk-cli==1.5
cycler==0.12.1
dask==2023.5.0
dask-image==2023.3.0
debugpy==1.8.5
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
dicomweb-client==0.59.3
diskcache==5.6.3
distlib==0.3.8
distributed==2023.5.0
dm-tree==0.1.8
docrep==0.3.2
et-xmlfile==1.1.0
etils==1.3.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
flax==0.7.2
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
get-annotations==0.1.2
girder-client==3.2.6
girder-slicer-cli-web==1.5.11
Glymur==0.13.6
google-auth==2.35.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
histomicstk==1.3.14
httpcore==1.0.5
httpx==0.27.2
igraph==0.11.6
imagecodecs==2023.3.16
imagecodecs-numcodecs==2023.3.16
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.13.1
kiwisolver==1.4.7
kornia_rs==0.1.8
large-image==1.30.3
large-image-converter==1.30.3
large-image-source-bioformats==1.30.3
large-image-source-deepzoom==1.30.3
large-image-source-dicom==1.30.3
large-image-source-dummy==1.30.3
large-image-source-gdal==1.30.3
large-image-source-mapnik==1.30.3
large-image-source-multi==1.30.3
large-image-source-nd2==1.30.3
large-image-source-ometiff==1.30.3
large-image-source-openjpeg==1.30.3
large-image-source-openslide==1.30.3
large-image-source-pil==1.30.3
large-image-source-rasterio==1.30.3
large-image-source-test==1.30.3
large-image-source-tiff==1.30.3
large-image-source-tifffile==1.30.3
large-image-source-vips==1.30.3
large-image-source-zarr==1.30.3
lazy_loader==0.4
leidenalg==0.10.2
libclang==18.1.1
lightning-utilities==0.11.7
llvmlite==0.38.1
locket==1.0.0
lpips==0.1.4
mapnik==4.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mdurl==0.1.2
mistune==3.0.2
ml-collections==0.1.1
ml-dtypes==0.2.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mudata==0.2.3
multidict==6.1.0
multipledispatch==1.0.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nd2==0.10.1
nest-asyncio==1.6.0
networkx==3.1
nimfa==1.4.0
notebook==7.2.2
notebook_shim==0.2.4
numba==0.55.2
numcodecs==0.12.1
numpy==1.22.4
numpyro==0.12.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
ome-types==0.5.3
omnipath==1.0.8
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
openslide-bin==4.0.0.6
openslide-python==1.4.1.1
optax==0.1.8
orbax-checkpoint==0.2.3
overrides==7.7.0
packaging==24.1
palettable==3.3.3
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.1
patsy==0.5.6
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare==0.7.5
pillow==10.4.0
PIMS==0.7
pkgconfig==1.5.5
pkgutil_resolve_name==1.3.10
platformdirs==4.3.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.2
pydantic-compat==0.1.2
pydantic_core==2.27.1
pydicom==2.4.4
Pygments==2.18.0
pylibtiff==0.6.1.1
pynndescent==0.5.13
pyparsing==3.1.4
pyproj==3.5.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-bioformats==4.1.0.8.0.1
python-dateutil==2.9.0.post0
python-igraph==0.11.6
python-javabridge==4.0.4
python-json-logger==2.0.7
pytorch-lightning==1.9.5
pytz==2024.2
pyvips==2.2.3
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.3.10
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
resource-backed-dask-array==0.1.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.1
rpds-py==0.20.0
s3transfer==0.10.4
scanpy==1.9.8
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scvi-tools==0.20.3
seaborn==0.13.2
Send2Trash==1.8.3
session-info==1.0.0
shapely==2.0.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slicerator==1.1.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
squidpy==1.2.3
stack-data==0.6.3
staintools==2.1.2
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.13.3
tblib==3.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorstore==0.1.45
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tifftools==1.5.2
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typeguard==4.3.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
uri-template==1.3.0
urllib3==1.26.20
validators==0.34.0
virtualenv==20.26.5
wcwidth==0.2.13
webcolors==24.8.0
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsidicom==0.13.0
xarray==2023.1.0
xsdata==24.3.1
yarl==1.11.1
zarr==2.16.1
zict==3.0.0
zipp==3.20.1
